CTU, PictureEncoder

+++ source/H265Lib/Common/FileIO.cpp
@@ -78,9 +78,9 @@ namespace HEVC
 				result = std::make_unique<Matrix<T>>(height, width);
 
 				T tmp;
-				for (int i = 0; i < height; ++i)
+				for (size_t i = 0; i < height; ++i)
 				{
-					for (int j = 0; j < width; ++j)
+					for (size_t j = 0; j < width; ++j)
 					{
 						str >> tmp;
 						if (str.eof() || str.fail())
